rewrote routing logic
